+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 ### Fixed:
 - Issue #115: drpall does not get updated when a tool sets a custom release.
 - Issue #107: missing os library under save function of Map class
+- Issue #117: hybrid colours were incorrect as they were being derived from petroth50_el.
 
 
 ## [2.0.9] - 2016/11/19